fix nova::network::quantum parameters
this patch updates the parameters that are passed to
the nova::network::quantum class for both the compute
and nova controller.

The most significant change is that references to quantum_connection_host
are removed b/c this parameter has not been available in nova since essex.

Other changes, remove commented out parameters, as well as updated unit tests.

Change-Id: I875c66fe881a31caa50dc10c6df72c4a74acdacc