ensure cinder volume_group sets the created volume
The openstack::cinder::storage class previously accepted
a parameter called volume_group which was used to set
the name of the volume group that cinder should use to
create managed volumes.

The same class also takes an argument called setup_test_volume
that indicates if a test_volume should be created. The problem
was that this test volume name was not being set as
$volume_group, meaning that they would not match up if the user
supplied a custom volume name.

This patch resolves that issue, ensuring that volume groups
are created with the correct name.

Change-Id: Icbd43ec35b2cc008e5fa84e717fd5804fbe280d0