Add swift_public_address
Previously, there was no way to specify
a separate public address for swift's endpoint
in keystone's service catalog.

This means that there was no way to support
installing swift proxy on a node besides the controller.

This commit adds a config param swift_public_address
that can be used to configure swift on a separate node.

Change-Id: Iafc415f69b85a97c45f6410be1d4888c61c39bba