Adds Support for Bind Host
Previously, the openstack::glance class would not allow
users to specify an IP address to bind Glance service to.

This change implements the bind_host parameter to allow
users to specify an IP address for binding Glance API and Registry.

Defaults to '0.0.0.0' for backwards compatibility and to bind
Glance services to all addresses.

Change-Id: I1e69df9401b2358b6af531c083850258c3585e41