Adds Support for Registry Host
Previously, the openstack::glance class would not allow
users to specify an IP address for API to find Registry.

This change implements the registry_host parameter to allow
users to specify an IP address for API to find the Registry service.

Defaults to '0.0.0.0' for backwards compatibility and allow all
addresses.

Change-Id: I2b330508eb849170ce958c5ed72ce183cf56c38c