Add Support for Separating Memcached from Horizon
Previously, the cache_server_ip parameter was used by the
memcached and horizon classes.  This prevented users from
binding memcached to a local address and having Horizon
use an external memcached server.

This change implements the memcached_listen_ip parameter, which
is used by the memcached class to allow users to seperate the IP
address used for binding memcached from the addressed used by
Horizon to access memcached.

Defaults to '127.0.0.1' for backwards compatibility and to have
memcached bind to a local address.

Change-Id: Ic4619335c59217f4759460b0f2dbad07b8a625c5