Seperate Quantum Host From Keystone Host
Previously, the nova::network::quantum class would use the
keystone_host parameter for the quantum_url.  This assumes that
the Quantum API will be running on the same host as Keystone.

Additoinally, the quantum_url parameter is used to specify the URL
for Nova to communicate with the Quantum API endpoint and not the
Keystone endpoint.

A new paramter called quantum_host has been introduced to
specify the IP address of the host running the Quantum API service.

Defaults to '127.0.0.1' for backwards compatability and to use the
local host for having Nova communicate with Quantum.

Change-Id: I7efface99f4be47bbb9262ec77285b843d60de15