Adds Support for Bind and Registry Hosts
Previously, the openstack::glance class would not allow
users to specify an IP address for API to find Registry.

Additionally, the class would not allow users to specify an
IP address for binding the Glance API and Registry services.

This change implements the registry_host parameter to allow
users to specify an IP address for API to find the Registry service.

Additionally, this change implements the bind_host parameter to allow
users to specify an IP address for binding Glance API and Registry.

bind_host defaults to '0.0.0.0' and registry_host defaults to
'127.0.0.1' for backwards compatibility.

Change-Id: Id309b1698bd6beca582bf01f8f4f7337f5ba5981