update rabbitmq reference
This commit updates the ref for rabbit to an earlier revision

While the previous revision worked, it was always enabling the
rabbitmq upstream repos causing the latest version of it to be
installed.

For openstack, we need to use the version supporting by the distros
(and tested with openstack) and not the latest version.

Change-Id: I8f81071929604271c4370019ea15ec7d14ee93fa