Adds Support for a Configuring Memcached
Previously, the memcached listen ip was hard-coded to '127.0.0.1'.
This does not allow users to bind memcached to a routable IP
address.  This is needed to allow memcached to service remote
requests.

The change introduces the memcached_listen_ip and
swift_memcache_servers parameters.  memcached_listen_ip allows
users to specify an IP address for memcached to listen on.
swift_memcache_servers specifies a list (ip:port) of memcached
servers that can be used by the Swift proxy cache service.

Defaults to '127.0.0.1' for backwards compatibility and to have
memcached only listen to local requests (loopback address).

Change-Id: I4287b9a165b5aae4568334b7b97f1bf40696c2f5