Exposes the Glance registry_host Parameter
Previously, the registry_host parameter was not exposed to the
openstack::controller class.  This prevented the class from
managing the parameter.

The new parameter glance_registry_host was added to allow users
to manage the IP address used by the Glance API service for
communicating with the Glance Registry service.

Defaults to '0.0.0.0' for backwards compatibility.

Change-Id: Ica07c59ab6ad02a825f67abd6bde5f1043993ac3