Proper selinux logic for Horizon
According to here:
http://docs.puppetlabs.com/learning/variables.html
"all facts are strings, and all non-empty strings,
 including the string "false" . are true"

Checking "if $::selinux" will always return true,
just like (if 'hello':) in python will always return true.

According to the above mentioned link, surronding the variable
in double quotes allows it to be interpreted as a string,
not a boolean.

This is also how RedHat does it, as seen below:

https://github.com/jsomara/astapor/blob/master/puppet/quickstack/manifests/controller.pp#L162-L167
https://github.com/stackforge/packstack/blob/master/packstack/puppet/templates/horizon.pp#L31-L36

Change-Id: I9489c3bd9a217c80762f5875f026676268572d79