Fixes Quantum ovs_local_ip Error Message
Previously, the openstack::quantum class would give an error
message when enable_ovs_agent was true and no local_ip was set.
This made sense when the module only supported the OVS GRE
networking model. Now that the module supports 'vlan' and 'gre'
tenant network types, this error is no longer accurate.

The error has been updated to be trigged when tenant_network_type
is 'gre' and no local_ip parameter is provided.

Change-Id: I4ba9c48b44b3733b44b7ba9c7e863ca74892feb5