Fix Swift Keystone Authentication Endpoints
Previously, the swift admin and internal auth endpoints would
default to internal_address.  This causes problems where a
different IP address is used for the swift public address
and a controller address is being used for the internal_address.
In this situation, swift admin/internal auth endpoints use the
controller address, instead of the swift public address.

Based on patch from master:
  https://review.openstack.org/#/c/40287/

Change-Id: I903748bf390180c4db9fff4fcb626727f6feb41f