Fix puppet-lint failures
This commit does two things to resolve
puppet-lint failures:

1. Split device_endpoint into its own file

Previously, the device_endpoint class was
defined inside of the storage-node.pp class.

This resulted in puppet-lint failures.

This commit moves the define to its own
file in order to conform to module
standards.

2. Modify spacing of a line in openstack::provision

This second change is actually a limitation in puppet-lint,
I have file: https://github.com/rodjek/puppet-lint/issues/213
to track it.

Change-Id: If83eb7fb740c5632711da6ab0f91bf6220c5830e