Add support for rbd-backed volumes to ::all class
The openstack::all class currently only supports the iscsi
backend for Cinder.  This commit allows users to specify
that Cinder volumes should be backed by Ceph RBD instead
by specifying a different volume driver and other necessary
Ceph information.

This commit is based on the following change to master but is
not cherry-picked due to the quantum->neutron name change mandating
a few changes to the rspec tests:
https://review.openstack.org/#/c/47310/

Change-Id: Ia4314d73e9e458f0be04d98fa0f50ff06d7cba0f
Implements: blueprint ceph-all-in-one