Enable RBD-backed Glance in openstack::all
Although openstack::glance supports using Ceph RBD as a backend,
the openstack::all class currently lacks parameters to enable
RBD-backed-glance in all-in-one installations.  This commit
adds the necessary parameters for users to use Ceph as a backend
for Glance.

This commit is based on the following change to master but is
not cherry-picked due to the quantum->neutron name change mandating
a few changes to the rspec tests:
https://review.openstack.org/#/c/47520/

Change-Id: I780e98fa28600d88607194fc0b4d82f26d72d063
Implements: blueprint rbd-backed-glance-in-all