Add force_config_drive to openstack::all
Currently the openstack::all class doesn't expose a parameter for the
force_config_drive option in nova.conf.  The underlying ::nova::compute
class does have such a parameter.  This commit exposes
force_config_drive as a parameter to the openstack::all class and passes
it down to the ::nova::compute class appropriately.  By default it is
set to false to preserve backward compatibility.

Change-Id: Iab0f0b1f358e22bfe2e79cf9c9adbd83f77eef7a
Implements: blueprint force-config-drive-for-all-class