Add force_config_drive to openstack::all
Currently the openstack::all class doesn't expose a parameter for the
force_config_drive option in nova.conf.  The underlying ::nova::compute
class does have such a parameter.  This commit exposes
force_config_drive as a parameter to the openstack::all class and passes
it down to the ::nova::compute class appropriately.  By default it is
set to false to preserve backward compatibility.

This patch is based on the following change to master , but is not
cherry picked due to the quantum->neutron name change requiring that
the rspec test be modified slightly:
https://review.openstack.org/#/c/47905/

Change-Id: I57a87123d02a2bbb748f466385e46bd68cd8bfbb
Implements: blueprint force-config-drive-for-all-class