Fix missing nova params for iscsi in ::all class
The openstack::all class fails to write three parameters to
nova.conf when used with iscsi Cinder volumes.  As a result,
although Cinder gets set up properly and volumes can be created,
attempts to attach the volumes to instances fail.  This patch
writes volume_group and iscsi_ip_address to nova.conf using the
same values that are already being provided to cinder.conf, and
adds a new parameter to the class for volume_name_template which
is also written to nova.conf.

Change-Id: I8342bb456b3827a8d8d22e06312743fd04f1fa60
Closes-Bug: #1237778