Module stackforge/puppet-horizon uses puppet templating to create the /etc/openstack-horizon/local_settings.py configuration file. That module also allows the end user manifests of the module to specify a different template via the parameter local_settings_template.
However, this module (stackforge/puppet-openstack) which uses the
puppet-horizon module does not expose this parameter. This creates
 somewhat of a problem as the default template sets ALLOWED_HOSTS to
the controller node. Not necessarily bad, unless one wants access from
other nodes. If so there is no way to apply a custom template.

This changes exposes the parameter and allows manifests using
puppet-openstack/controller to pass in a custom local_settings template
path.

Change-Id: Ibfc8e4db01ddf04527ce7ff7513e5e1afdba4a9c