Add parameter for custom Horizon local settings template
Module stackforge/puppet-horizon uses puppet templating to create the
/etc/openstack-horizon/local_settings.py configuration file. That module
also allows the end user manifests of the module to specify a different template
via the parameter local_settings_template.

However, this module (stackforge/puppet-openstack) which uses the
puppet-horizon module does not expose this parameter. So there is no way to apply a custom template.

This change exposes the parameter and allows manifests using
puppet-openstack/controller or puppet-openstack/all to pass in a custom local_settings template path.

Change-Id: Ibfc8e4db01ddf04527ce7ff7513e5e1afdba4a9c